Refactored the PolarsTableModel class and generate_statistics function to separate files.
